# docs/OpenAPI.yaml
openapi: 3.0.3
info:
  title: FanCollector API
  description: API REST para gestionar colecciones de fans. Permite a los usuarios crear y administrar colecciones de items coleccionables como photocards, albums, merchandising, etc.
  version: 1.0.0
  contact:
    name: NereaTM
    email: admin@fancollector.com
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo local
...
components:
  schemas:
...
    UsuarioColeccionInDTO:
      type: object
      required:
        - idUsuario
        - idColeccion
      properties:
        idUsuario:
          type: integer
          format: int64
          example: 1
        idColeccion:
          type: integer
          format: int64
          example: 2
        esFavorita:
          type: boolean
          default: false
          example: true
        esCreador:
          type: boolean
          default: false
          example: false
        esVisible:
          type: boolean
          default: true
          example: true
...
    UsuarioPutDTO:
      type: object
      required:
        - nombre
        - email
      properties:
        nombre:
          type: string
          minLength: 1
          example: Nerea Updated
        email:
          type: string
          format: email
          example: nerea.updated@gmail.com
        urlAvatar:
          type: string
          example: ejemplo/avatar-nuevo.jpg
        descripcion:
          type: string
          example: Fan de BTS, anime y Pokemon
        contactoPublico:
          type: string
          example: "@nereanew"
        archivo:
          type: string
          format: binary
          description: Imagen (jpg, png, gif, máx 15MB)
...
    UsuarioColeccionPutDTO:
      type: object
      properties:
        esFavorita:
          type: boolean
          example: false
        esCreador:
          type: boolean
          example: true
        esVisible:
          type: boolean
          example: true
...
    UsuarioOutDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nombre:
          type: string
          example: Nerea Test
        email:
          type: string
          format: email
          example: nerea.test@gmail.com
        rol:
          $ref: '#/components/schemas/RolUsuario'
        urlAvatar:
          type: string
          example: ejemplo/avatar.jpg
        descripcion:
          type: string
          example: Fan de BTS y anime
        contactoPublico:
          type: string
          example: "@nereatest"
        fechaRegistro:
          type: string
          format: date-time
          example: "2025-12-01T10:30:00"
...
    UsuarioColeccionOutDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 3
        idUsuario:
          type: integer
          format: int64
          example: 1
        idColeccion:
          type: integer
          format: int64
          example: 2
        esFavorita:
          type: boolean
          example: true
        esCreador:
          type: boolean
          example: false
        esVisible:
          type: boolean
          example: true
        fechaAgregada:
          type: string
          format: date-time
          example: "2025-12-05T14:20:00"
...
    ErrorRespuesta:
      type: object
      properties:
        statusCode:
          type: integer
          example: 404
        message:
          type: string
          example: Usuario con id 999 no encontrado
        errors:
          type: object
          additionalProperties:
            type: string
          example:
            nombre: "El nombre no puede estar en blanco"
            email: "El email debe tener un formato valido"
paths:
...
  /usuarios/{id}:
    get:
      tags:
        - Usuarios
      summary: Obtener un usuario por ID
      description: Recupera la informacion de un usuario especifico
      parameters:
        - name: id
          in: path
          required: true
          description: ID del usuario
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioOutDTO'
              examples:
                usuarioEncontrado:
                  summary: Usuario existente
                  value:
                    id: 1
                    nombre: Nerea Test
                    email: nerea.test@gmail.com
                    rol: USER
                    urlAvatar: ejemplo/nerea.jpg
                    descripcion: Fan de BTS y anime
                    contactoPublico: "@nereatest"
                    fechaRegistro: "2025-12-01T10:30:00"
                usuarioAdmin:
                  summary: Usuario administrador
                  value:
                    id: 2
                    nombre: Juan Test
                    email: juan.test@gmail.com
                    rol: ADMIN
                    urlAvatar: ejemplo/juan.jpg
                    descripcion: Coleccionista de Pokemon
                    contactoPublico: "@juantest"
                    fechaRegistro: "2025-11-15T08:20:00"
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 404
                message: Usuario con id 999 no encontrado
                errors: { }
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 500
                message: Error interno del servidor
                errors: { }

    put:
      tags:
        - Usuarios
      summary: Actualizar un usuario
      description: Actualiza la informacion de un usuario existente
      parameters:
        - name: id
          in: path
          required: true
          description: ID del usuario
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UsuarioPutDTO'
            examples:
              actualizacionBasica:
                summary: Actualizacion basica
                value:
                  nombre: Nerea Updated
                  email: nerea.updated@gmail.com
                  contrasena: newPassword123
              actualizacionCompleta:
                summary: Actualizacion completa
                value:
                  nombre: Juan Updated Test
                  email: juan.updated@gmail.com
                  contrasena: newSecurePass789
                  urlAvatar: ejemplo/juan-new.jpg
                  descripcion: Coleccionista experto de Pokemon y Yu-Gi-Oh
                  contactoPublico: "@juanupdated"
      responses:
        '200':
          description: Usuario actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioOutDTO'
              examples:
                usuarioActualizado:
                  summary: Usuario actualizado
                  value:
                    id: 1
                    nombre: Nerea Updated
                    email: nerea.updated@gmail.com
                    rol: USER
                    urlAvatar: ejemplo/nerea.jpg
                    descripcion: Fan de BTS y anime
                    contactoPublico: "@nereatest"
                    fechaRegistro: "2025-12-01T10:30:00"
                usuarioCompletoActualizado:
                  summary: Usuario completamente actualizado
                  value:
                    id: 2
                    nombre: Juan Updated Test
                    email: juan.updated@gmail.com
                    rol: ADMIN
                    urlAvatar: ejemplo/juan-new.jpg
                    descripcion: Coleccionista experto de Pokemon y Yu-Gi-Oh
                    contactoPublico: "@juanupdated"
                    fechaRegistro: "2025-11-15T08:20:00"
        '400':
          description: Datos de entrada invalidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 400
                message: Errores de validacion
                errors:
                  email: El email debe tener un formato valido
                  contrasena: La contrasena debe tener al menos 4 caracteres
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 404
                message: Usuario con id 999 no encontrado
                errors: { }
        '403':
          description: Acceso denegado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 403
                message: No tienes permisos para realizar esta acción
                errors: { }
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 500
                message: Error interno del servidor
                errors: { }

    delete:
      tags:
        - Usuarios
      summary: Eliminar un usuario
      description: Elimina un usuario del sistema
      parameters:
        - name: id
          in: path
          required: true
          description: ID del usuario
          schema:
            type: integer
            format: int64
          example: 4
      responses:
        '204':
          description: Usuario eliminado exitosamente
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 404
                message: Usuario con id 999 no encontrado
                errors: { }
        '403':
          description: Acceso denegado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 403
                message: No tienes permisos para realizar esta acción
                errors: { }
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 500
                message: Error interno del servidor
                errors: { }
...
  /usuario-colecciones:
    post:
      tags:
        - Usuario-Colecciones
      summary: Crear relacion usuario-coleccion
      description: Asocia un usuario con una coleccion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioColeccionInDTO'
            examples:
              relacionBasica:
                summary: Relacion basica
                value:
                  idUsuario: 1
                  idColeccion: 2
                  esFavorita: true
                  esCreador: false
              relacionCreador:
                summary: Usuario es creador
                value:
                  idUsuario: 2
                  idColeccion: 3
                  esFavorita: true
                  esCreador: true
      responses:
        '201':
          description: Relacion creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioColeccionOutDTO'
              examples:
                relacionCreada:
                  summary: Relacion recien creada
                  value:
                    id: 3
                    idUsuario: 1
                    idColeccion: 2
                    esFavorita: true
                    esCreador: false
                    fechaAgregada: "2025-12-13T17:30:00"
                relacionCreadorCreada:
                  summary: Relacion creador creada
                  value:
                    id: 4
                    idUsuario: 2
                    idColeccion: 3
                    esFavorita: true
                    esCreador: true
                    fechaAgregada: "2025-12-13T18:00:00"
        '400':
          description: Datos invalidos o relacion ya existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              examples:
                validacionError:
                  summary: Errores de validacion
                  value:
                    statusCode: 400
                    message: Errores de validacion
                    errors:
                      idUsuario: El ID del usuario es obligatorio
                      idColeccion: El ID de la coleccion es obligatorio
                relacionDuplicada:
                  summary: Relacion ya existe
                  value:
                    statusCode: 400
                    message: La relacion ya existe
                    errors: { }
        '404':
          description: Usuario o coleccion no encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              examples:
                usuarioNoEncontrado:
                  summary: Usuario no existe
                  value:
                    statusCode: 404
                    message: Usuario con id 999 no encontrado
                    errors: { }
                coleccionNoEncontrada:
                  summary: Coleccion no existe
                  value:
                    statusCode: 404
                    message: Coleccion con id 999 no encontrada
                    errors: { }
        '403':
          description: Acceso denegado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 403
                message: No tienes permisos para realizar esta acción
                errors: { }
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 500
                message: Error interno del servidor
                errors: { }

    get:
      tags:
        - Usuario-Colecciones
      summary: Listar relaciones usuario-coleccion con filtros opcionales
      description: Obtiene lista de relaciones entre usuarios y colecciones. Se pueden aplicar filtros.
      parameters:
        - name: idUsuario
          in: query
          description: Filtrar por ID de usuario
          required: false
          schema:
            type: integer
            format: int64
          example: 1
        - name: idColeccion
          in: query
          description: Filtrar por ID de coleccion
          required: false
          schema:
            type: integer
            format: int64
          example: 2
        - name: soloFavoritas
          in: query
          description: Filtrar solo colecciones favoritas
          required: false
          schema:
            type: boolean
          example: true
        - name: esVisible
          in: query
          description: Filtrar por visibilidad
          required: false
          schema:
            type: boolean
          example: true
      responses:
        '200':
          description: Lista de relaciones obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsuarioColeccionOutDTO'
              examples:
                listaRelaciones:
                  summary: Lista con dos relaciones
                  value:
                    - id: 1
                      idUsuario: 1
                      idColeccion: 1
                      esFavorita: true
                      esCreador: true
                      esVisible: true
                      fechaAgregada: "2025-12-03T13:15:00"
                    - id: 2
                      idUsuario: 1
                      idColeccion: 2
                      esFavorita: false
                      esCreador: false
                      esVisible: true
                      fechaAgregada: "2025-12-05T14:20:00"
                soloFavoritas:
                  summary: Solo colecciones favoritas
                  value:
                    - id: 1
                      idUsuario: 1
                      idColeccion: 1
                      esFavorita: true
                      esCreador: true
                      esVisible: true
                      fechaAgregada: "2025-12-03T13:15:00"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 500
                message: Error interno del servidor
                errors: { }
  /usuario-colecciones/{id}:
    get:
      tags:
        - Usuario-Colecciones
      summary: Obtener relacion usuario-coleccion por ID
      description: Recupera una relacion especifica
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la relacion
          schema:
            type: integer
            format: int64
          example: 3
      responses:
        '200':
          description: Relacion encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioColeccionOutDTO'
              examples:
                relacionEncontrada:
                  summary: Relacion existente
                  value:
                    id: 3
                    idUsuario: 1
                    idColeccion: 2
                    esFavorita: true
                    esCreador: false
                    esVisible: true
                    fechaAgregada: "2025-12-05T14:20:00"
                relacionCreador:
                  summary: Relacion de creador
                  value:
                    id: 1
                    idUsuario: 1
                    idColeccion: 1
                    esFavorita: true
                    esCreador: true
                    esVisible: true
                    fechaAgregada: "2025-12-03T13:15:00"
        '404':
          description: Relacion no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 404
                message:Usuario-coleccion con ID: 999 no encontrado
                errors: { }
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 500
                message: Error interno del servidor
                errors: { }

    put:
      tags:
        - Usuario-Colecciones
      summary: Actualizar relacion usuario-coleccion
      description: Actualiza una relacion existente
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la relacion
          schema:
            type: integer
            format: int64
          example: 2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioColeccionPutDTO'
            examples:
              marcarFavorita:
                summary: Marcar como favorita
                value:
                  esFavorita: true
                  esCreador: false
                  esVisible: true
              cambiarCreador:
                summary: Cambiar estado de creador
                value:
                  esFavorita: true
                  esCreador: true
                  esVisible: true
      responses:
        '200':
          description: Relacion actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioColeccionOutDTO'
              examples:
                relacionActualizada:
                  summary: Relacion actualizada
                  value:
                    id: 2
                    idUsuario: 1
                    idColeccion: 2
                    esFavorita: true
                    esCreador: false
                    esVisible: true
                    fechaAgregada: "2025-12-05T14:20:00"
        '404':
          description: Relacion no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 404
                message:Usuario-coleccion con ID: 999 no encontrado
                errors: { }
        '403':
          description: Acceso denegado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 403
                message: No tienes permisos para realizar esta acción
                errors: { }
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 500
                message: Error interno del servidor
                errors: { }

    delete:
      tags:
        - Usuario-Colecciones
      summary: Eliminar relacion usuario-coleccion (1)
      deprecated: true
      description: U sar DELETE /usuario-colecciones/v2/{id}. Elimina una relacion entre usuario y coleccion pero los items en relacion con la coleccion no se eliminan
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la relacion
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '204':
          description: Relacion eliminada exitosamente
        '404':
          description: Relacion no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 404
                message:Usuario-coleccion con ID: 999 no encontrado
                errors: { }
        '403':
          description: Acceso denegado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 403
                message: No tienes permisos para realizar esta acción
                errors: { }
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 500
                message: Error interno del servidor
                errors: { }
...
  /usuario-colecciones/{id}/favorita:
    patch:
      tags:
        - Usuario-Colecciones
      summary: Actualizar estado de favorita
      description: Cambia si una coleccion es favorita para el usuario
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la relacion
          schema:
            type: integer
            format: int64
          example: 2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioColeccionFavoritaDTO'
            examples:
              marcarFavorita:
                summary: Marcar como favorita
                value:
                  esFavorita: true
              desmarcarFavorita:
                summary: Desmarcar como favorita
                value:
                  esFavorita: false
      responses:
        '200':
          description: Estado de favorita actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioColeccionOutDTO'
              examples:
                ahoraFavorita:
                  summary: Coleccion ahora favorita
                  value:
                    id: 2
                    idUsuario: 1
                    idColeccion: 2
                    esFavorita: true
                    esCreador: false
                    esVisible: true
                    fechaAgregada: "2025-12-05T14:20:00"
                yaNoFavorita:
                  summary: Coleccion ya no favorita
                  value:
                    id: 2
                    idUsuario: 1
                    idColeccion: 2
                    esFavorita: false
                    esCreador: false
                    fechaAgregada: "2025-12-05T14:20:00"
        '400':
          description: Valor invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 400
                message: Errores de validacion
                errors:
                  esFavorita: Favorita no puede ser nulo
        '404':
          description: Relacion no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 404
                message:Usuario-coleccion con ID: 999 no encontrado
                errors: { }
        '403':
          description: Acceso denegado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 403
                message: No tienes permisos para realizar esta acción
                errors: { }
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 500
                message: Error interno del servidor
                errors: { }
  /usuario-colecciones/{id}/visible:
    patch:
      tags:
        - Usuario-Colecciones
      summary: Actualizar visibilidad de coleccion en perfil
      description: Cambia si una coleccion agregada es visible en el perfil del usuario
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la relacion
          schema:
            type: integer
            format: int64
          example: 2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioColeccionVisibleDTO'
            examples:
              hacerVisible:
                summary: Hacer coleccion visible en perfil
                value:
                  esVisible: true
              ocultarColeccion:
                summary: Ocultar coleccion del perfil
                value:
                  esVisible: false
      responses:
        '200':
          description: Visibilidad actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioColeccionOutDTO'
              examples:
                coleccionVisible:
                  summary: Coleccion ahora visible
                  value:
                    id: 2
                    idUsuario: 1
                    idColeccion: 2
                    esFavorita: true
                    esCreador: false
                    esVisible: true
                    fechaAgregada: "2025-12-05T14:20:00"
                coleccionOculta:
                  summary: Coleccion ahora oculta
                  value:
                    id: 2
                    idUsuario: 1
                    idColeccion: 2
                    esFavorita: true
                    esCreador: false
                    esVisible: false
                    fechaAgregada: "2025-12-05T14:20:00"
        '400':
          description: Valor invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 400
                message: Errores de validacion
                errors:
                  esVisible: El campo esVisible no puede ser nulo
        '404':
          description: Relacion no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 404
                message:
                  Usuario-coleccion con ID: 999 no encontrado
                errors: { }
        '403':
          description: Acceso denegado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 403
                message: No tienes permisos para realizar esta acción
                errors: { }
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                statusCode: 500
                message: Error interno del servidor
                errors: { }
  # USUARIO-ITEMS
...